Move the main method to class HomeWork03.java
