+++ src/com/softserve/edu10/hm10/Student.java
@@ -20,6 +20,15 @@ class Student implements Comparable<Student> {
     }
 
     //Move this method in class where is main method
+
+
+    @Override
+    public int compareTo(Student other) {
+        return this.name.compareTo(other.name);
+    }
+
+    public static Comparator<Student> compareByCourse = Comparator.comparingInt(student -> student.course);
+
     public void printStudents(List<Student> students, int courseNumber) {
         System.out.println("Students enrolled in course " + courseNumber + ":");
         for (Student student : students) {
@@ -29,40 +38,4 @@ class Student implements Comparable<Student> {
         }
     }
 
-    @Override
-    public int compareTo(Student other) {
-        return this.name.compareTo(other.name);
-    }
-
-    //This variable is good
-    public static Comparator<Student> compareByCourse = Comparator.comparingInt(student -> student.course);
-
-    //Move this main method to class like App.java
-    public static void main(String[] args) {
-        List<Student> students = new ArrayList<>();
-        students.add(new Student("John", 1));
-        students.add(new Student("Alice", 2));
-        students.add(new Student("Bob", 1));
-        students.add(new Student("Charlie", 3));
-        students.add(new Student("David", 2));
-
-        // Display list of students sorted by name
-        System.out.println("Students sorted by name:");
-        Collections.sort(students);
-        for (Student student : students) {
-            System.out.println(student.getName() + " - Course: " + student.getCourse());
-        }
-
-        // Display list of students sorted by course
-        System.out.println("\nStudents sorted by course:");
-        students.sort(compareByCourse);
-        for (Student student : students) {
-            System.out.println(student.getName() + " - Course: " + student.getCourse());
-        }
-
-        // Print students enrolled in course 2
-        System.out.println("\nStudents enrolled in course 2:");
-        Student studentObj = new Student("", 0);
-        studentObj.printStudents(students, 2);
-    }
 }
